requirements.txt
py-cord==2.6.1
requests==2.32.3
python-dotenv==1.0.1
dacite==1.9.2
APScheduler==3.11.0
psycopg2-binary==2.9.11
